+++ DQM/HcalTasks/interface/HcalHarvesting.h
